Standardise float64 -> uint64 conversions on macOS/ARM
See https://github.com/golang/go/issues/52943

Go Spec says "In all non-constant conversions involving floating-point
or complex values, if the result type cannot represent the value the
conversion succeeds but the result value is implementation-dependent."

On x64 this does uint64(int64(f)), but on ARM it yields 0. The former
behaviour seems more intuitive, so run with that.

+++ interp/functions.go
@@ -82,15 +82,15 @@ func (p *interp) toNative(v value, typ reflect.Type) reflect.Value {
 	case reflect.Int64:
 		return reflect.ValueOf(int64(v.num()))
 	case reflect.Uint:
-		return reflect.ValueOf(uint(v.num()))
+		return reflect.ValueOf(uint(int64(v.num())))
 	case reflect.Uint8:
-		return reflect.ValueOf(uint8(v.num()))
+		return reflect.ValueOf(uint8(int64(v.num())))
 	case reflect.Uint16:
-		return reflect.ValueOf(uint16(v.num()))
+		return reflect.ValueOf(uint16(int64(v.num())))
 	case reflect.Uint32:
-		return reflect.ValueOf(uint32(v.num()))
+		return reflect.ValueOf(uint32(int64(v.num())))
 	case reflect.Uint64:
-		return reflect.ValueOf(uint64(v.num()))
+		return reflect.ValueOf(uint64(int64(v.num())))
 	case reflect.Float32:
 		return reflect.ValueOf(float32(v.num()))
 	case reflect.Float64:
@@ -420,7 +420,7 @@ func (p *interp) sprintf(format string, args []value) (string, error) {
 		case 'f':
 			v = a.num()
 		case 'u':
-			v = uint64(a.num())
+			v = uint64(int64(a.num()))
 		case 'c':
 			var c []byte
 			n, isStr := a.isTrueStr()
